additional refactors
